changes to folders and scripts
moves screens related files info "screens" folder.
added scripts for easier local startup
