+++ templates/v0-clone/.env.example
@@ -5,9 +5,13 @@ NORTHFLANK_PROJECT_ID=vibe0
 
 CONVEX_DEPLOYMENT=
 
-NEXT_PUBLIC_CONVEX_URL=
 
 AUTH_SECRET=vibe0
 AUTH_GITHUB_ID=
 AUTH_GITHUB_SECRET=
+
+NEXT_PUBLIC_CONVEX_URL=
 NEXTAUTH_URL=http://localhost:3000
+NEXTAUTH_SECRET=
+
+INNGEST_EVENT_KEY=
