+++ templates/v0-clone/app/page.tsx
@@ -1,11 +1,1 @@
-import { Metadata } from "next";
-import ClientPage from "./client-page";
-
-export const metadata: Metadata = {
-  title: "Vibe0 | VibeKit",
-  description: "A v0 clone built with VibeKit",
-};
-
-export default function Home() {
-  return <ClientPage />;
-}
+export { default } from './screens/page';
