+++ templates/v0-clone/app/screens/page.tsx
@@ -1,0 +1,11 @@
+import { Metadata } from "next";
+import ClientPage from "./client-page";
+
+export const metadata: Metadata = {
+  title: "Vibe0 | VibeKit",
+  description: "A v0 clone built with VibeKit",
+};
+
+export default function Home() {
+  return <ClientPage />;
+}
