+++ templates/v0-clone/app/screens/session/[id]/page.tsx
@@ -1,0 +1,11 @@
+import ClientPage from "./client-page";
+
+interface Props {
+  params: Promise<{ id: string }>;
+}
+
+export default async function Home({ params }: Props) {
+  const { id } = await params;
+
+  return <ClientPage id={id} />;
+}
