+++ templates/v0-clone/app/screens/sessions/page.tsx
@@ -1,0 +1,11 @@
+import { Metadata } from "next";
+import SessionsClientPage from "./client-page";
+
+export const metadata: Metadata = {
+  title: "Sessions | vibe0",
+  description: "Sessions list",
+};
+
+export default async function SessionsPage() {
+  return <SessionsClientPage />;
+}
