+++ templates/v0-clone/app/session/[id]/page.tsx
@@ -1,11 +1,1 @@
-import ClientPage from "./client-page";
-
-interface Props {
-  params: Promise<{ id: string }>;
-}
-
-export default async function Home({ params }: Props) {
-  const { id } = await params;
-
-  return <ClientPage id={id} />;
-}
+export { default } from '../../screens/session/[id]/page';
