+++ templates/v0-clone/app/sessions/page.tsx
@@ -1,11 +1,1 @@
-import { Metadata } from "next";
-import SessionsClientPage from "./client-page";
-
-export const metadata: Metadata = {
-  title: "Sessions | vibe0",
-  description: "Sessions list",
-};
-
-export default async function SessionsPage() {
-  return <SessionsClientPage />;
-}
+export { default } from '../screens/sessions/page';
